refactor(publish): split postiz adapter into focused modules
Break the 673-line postiz.ts into client/markers/media/adapter/admin
under src/core/publish/postiz/, move listNativeIntegrations into
adapters.ts, and extract shared finalizeAdapterResult + dry-run helper.
